Extract RiffContainer
