+++ src/FileSignatures/FileFormat.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
+using FileSignatures.Extensions;
 
 namespace FileSignatures
 {
@@ -111,7 +111,7 @@ namespace FileSignatures
                 return false;
             }
 
-            return CheckContains(stream, Signature, Offset);
+            return stream.CheckContains(Signature, Offset);
         }
 
         /// <summary>
@@ -175,29 +175,6 @@ namespace FileSignatures
         public override string ToString()
         {
             return MediaType;
-        }
-
-        /// <summary>
-        /// Check stream contains sequence at specific offset.
-        /// </summary>
-        /// <param name="stream">The stream to check.</param>
-        /// <param name="sequence">The subsequence which should be found.</param>
-        /// <param name="offset">Offset in the stream at which the subsequence should be located.</param>
-        /// <returns></returns>
-        protected static bool CheckContains(Stream stream, IEnumerable<byte> sequence, int offset = 0)
-        {
-            stream.Position = offset;
-
-            foreach (var el in sequence)
-            {
-                var b = stream.ReadByte();
-                if (b != el)
-                {
-                    return false;
-                }
-            }
-
-            return true;
         }
     }
 }
